PARQUET-777: Update for review feedback.
